Added ability to filter trips based on the start stop.

+++ app/src/main/java/ro/go/stecker/stblogger/data/database/entities/Stop.kt
@@ -12,7 +12,7 @@ enum class StopType {
 @Entity
 data class Stop(
     @PrimaryKey(autoGenerate = false)
-    val id: Double = 0.0,
+    val id: Int = 0,
     val name: String = "",
     val description: String = "",
     val latitude: Double = 0.0,
